refactor to fix cycle import
